refactor: separate asset path resolution
